refactor: fix and clean up remaining resolvers
