+++ mycore-base/src/main/java/org/mycore/common/xsl/uriresolver/MCRExceptionAsXMLResolver.java
@@ -1,0 +1,71 @@
+/*
+ * This file is part of ***  M y C o R e  ***
+ * See https://www.mycore.de/ for details.
+ *
+ * MyCoRe is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MyCoRe is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MyCoRe.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.mycore.common.xsl.uriresolver;
+
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
+
+import javax.xml.transform.Source;
+import javax.xml.transform.URIResolver;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.jdom2.Element;
+import org.jdom2.Namespace;
+import org.jdom2.transform.JDOMSource;
+import org.mycore.common.MCRException;
+
+public class MCRExceptionAsXMLResolver implements URIResolver {
+
+    private static final Logger LOGGER = LogManager.getLogger();
+
+    @Override
+    public Source resolve(String href, String base) {
+        String target = href.substring(href.indexOf(':') + 1);
+
+        try {
+            return MCRURIResolver.obtainInstance().resolve(target, base);
+        } catch (Exception ex) {
+            LOGGER.debug("Caught {}. Put it into XML to process in XSL!", () -> ex.getClass().getName());
+            Element exception = new Element("exception");
+            Element message = new Element("message");
+            Element stacktraceElement = new Element("stacktrace");
+
+            stacktraceElement.setAttribute("space", "preserve", Namespace.XML_NAMESPACE);
+
+            exception.addContent(message);
+            exception.addContent(stacktraceElement);
+
+            message.setText(ex.getMessage());
+
+            try (StringWriter sw = new StringWriter(); PrintWriter pw = new PrintWriter(sw)) {
+                ex.printStackTrace(pw);
+                stacktraceElement.setText(pw.toString());
+            } catch (IOException e) {
+                MCRException mcrException = new MCRException("Error while writing Exception to String!", e);
+                mcrException.addSuppressed(ex);
+                throw mcrException;
+            }
+
+            return new JDOMSource(exception);
+        }
+    }
+
+}
